Use consistent new membership year logic
When a user created a new membership, it would be assigned to next year if the month is November or December. But UsersController#import, for imports of users who signed up in the tailgate lots, it was always just the current year. This centralizes the logic to behave the former way in both spots.