Initial Kotlin commit
Got ktor set up. Tried to get hot reloading working. I think it works. Hooking in the Intellij debugger is still not working despite the runConfigurations I've checked in. It says it's attached, but then it doesn't do anything when it should be encountering a breakpoint. So I don't know. Will futz with that as I continue setting up